refactor: consolidate 6 banner blocks into one showing all 4 statuses
Replace 6 individual blocks (Info, Action, Dismiss, Errors, Collapsible,
Full Width) with a single BannerStatuses block that shows info, success,
warning, and error side by side.

Made-with: Cursor
