--- packages/cli/templates/blocks/components/Banner/BannerCollapsibleContent.doc.mjs
@@ -1,9 +1,0 @@
-/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'block',
-  name: 'Banner — Collapsible',
-  description: 'Combine an action button, dismiss control, and expandable detail area in one banner. Use for complex notifications like config changes or deployment summaries.',
-  isReady: true,
-  aspectRatio: 16 / 4,
-  componentsUsed: ['Banner', 'Button', 'List', 'Layout', 'Text'],
-};
