--- packages/cli/templates/blocks/components/Banner/BannerDismissable.doc.mjs
@@ -1,9 +1,0 @@
-/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'block',
-  name: 'Banner — Dismiss',
-  description: 'Let the user close a banner after reading it. Use for maintenance notices, feature tips, or any non-critical message the user can acknowledge.',
-  isReady: true,
-  aspectRatio: 16 / 4,
-  componentsUsed: ['Banner', 'Layout'],
-};
