--- packages/cli/templates/blocks/components/Banner/BannerErrorList.doc.mjs
@@ -1,9 +1,0 @@
-/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'block',
-  name: 'Banner — Errors',
-  description: 'Show a list of errors inside an expandable banner. Use at the top of a form to summarize all validation problems at once.',
-  isReady: true,
-  aspectRatio: 16 / 4,
-  componentsUsed: ['Banner', 'Button', 'List'],
-};
