--- packages/cli/templates/blocks/components/Banner/BannerSectionVariant.doc.mjs
@@ -1,9 +1,0 @@
-/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'block',
-  name: 'Banner — Full Width',
-  description: 'A full-width banner with no border radius for page-level notifications. Use at the top of a page for site-wide announcements or maintenance alerts.',
-  isReady: true,
-  aspectRatio: 16 / 4,
-  componentsUsed: ['Banner', 'Layout'],
-};
