+++ packages/cli/templates/blocks/components/Banner/BannerStatuses.doc.mjs
@@ -1,0 +1,9 @@
+/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
+export const doc = {
+  type: 'block',
+  name: 'Banner — Statuses',
+  description: 'All 4 banner statuses: info, success, warning, and error. Use to show persistent messages like updates, confirmations, cautions, or problems at the top of a page or section.',
+  isReady: true,
+  aspectRatio: 16 / 9,
+  componentsUsed: ['Banner', 'Layout'],
+};
