+++ packages/cli/templates/blocks/components/Banner/BannerStatuses.tsx
@@ -3,13 +3,13 @@
 import {XDSBanner} from '@xds/core/Banner';
 import {XDSStack} from '@xds/core/Layout';
 
-export default function BannerWithDescription() {
+export default function BannerStatuses() {
   return (
     <XDSStack direction="vertical" gap={4}>
       <XDSBanner
         status="info"
         title="A new software update is available"
-        description="Version 2.4.1 includes performance improvements and bug fixes. Update at your convenience."
+        description="Version 2.4.1 includes performance improvements and bug fixes."
       />
       <XDSBanner
         status="success"
@@ -19,7 +19,12 @@ export default function BannerWithDescription() {
       <XDSBanner
         status="warning"
         title="Storage almost full"
-        description="You have used 90% of your available storage. Consider upgrading your plan."
+        description="You have used 90% of your available storage."
+      />
+      <XDSBanner
+        status="error"
+        title="Payment failed"
+        description="We could not process your last payment. Please update your billing details."
       />
     </XDSStack>
   );
