--- packages/cli/templates/blocks/components/Banner/BannerWithDescription.doc.mjs
@@ -1,9 +1,0 @@
-/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'block',
-  name: 'Banner — Info',
-  description: 'Show a message with a title and description. Use for system updates, feature announcements, or success confirmations.',
-  isReady: true,
-  aspectRatio: 16 / 4,
-  componentsUsed: ['Banner', 'Layout'],
-};
